+++ test/never-exit.ts
@@ -1,0 +1,3 @@
+import * as http from 'http'
+
+http.createServer(() => {}).listen()